Make it possible to add Attacker entry points
When a line is dragged to/from attacker item to/from asset item, the attacker
and asset item is identified and a dialog opens.

By checking in AttackerAttachment.entry_points we only list the attack steps
that are not already part of attacker entry points in the dialog.

When dialog is 'accepted', we add the entrypoint using
AttackerAttachment.add_entrypoint.